% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domaintools-package.r
\docType{package}
\name{domaintools}
\alias{domaintools}
\alias{domaintools-package}
\title{A R API interface to the DomainTools API}
\description{
A R API interface to the DomainTools API
}
\author{
Bob Rudis (@hrbrmstr)
}

