% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{domaintools_api_key}
\alias{domaintools_api_key}
\title{Get or set DOMAINTOOLS_API_KEY value}
\usage{
domaintools_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new DomainTools API key for the current environment?}
}
\value{
atomic character vector containing the DomainTools API key
}
\description{
The API wrapper functions in this package all rely on a DomainTools API key
in the environment variables \code{DOMAINTOOLS_API_KEY}. The easiest way to
accomplish this is to set it in the `.Renviron` file in your home directory.
}

