% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{domaintools_username}
\alias{domaintools_username}
\title{Get or set DOMAINTOOLS_API_USERNAME value}
\usage{
domaintools_username(force = FALSE)
}
\arguments{
\item{force}{Force setting a new DomainTools username for the current environment?}
}
\value{
atomic character vector containing the DomainTools username
}
\description{
The API wrapper functions in this package all rely on a DomainTools
username residing in the environment variables \code{DOMAINTOOLS_API_USERNAME}
The easiest way to accomplish this is to set it in the `.Renviron` file in your home directory.
}

