% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hosting_history.r
\name{hosting_history}
\alias{hosting_history}
\title{Hosting History}
\usage{
hosting_history(domain)
}
\arguments{
\item{domain}{domain to perform the Whois query on}
}
\value{
a \code{list} of hosting history result detais for the \code{domain}
}
\description{
The Hosting History API provides a list of changes that have
occurred in a Domain Name's registrar, IP address, and name servers.
IP and name server events include	the value before and after the change
and indicate the type of action that triggered the event.
}
\note{
In rare cases, you may request a domain for which no recent Whois
      record is available. If that occurs, the system will respond with
      an error.
}
\examples{
hosting_history("domaintools.com")
hosting_history("dailychanges.com")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/hosting-history/}
}

