% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reverse_ip.R
\name{reverse_ip}
\alias{reverse_ip}
\title{Reverse IP}
\usage{
reverse_ip(ip, limit = NULL)
}
\arguments{
\item{ip}{IP address to perform the reverse IP query}

\item{limit}{Limits the size of the domain list than can appear in a
response. The limit is applied per-IP address, not for the entire
request.}
}
\value{
a \code{list} of result detais for the \code{ip}
}
\description{
The Reverse IP API provides a list of domain names that share the same
Internet host (i.e. the same IP address).
}
\note{
In rare cases, you may request an IP for which no recent Whois
      record is available. If that occurs, the system will respond with
      an error.
}
\examples{
reverse_ip("64.246.165.240")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/reverse-ip/}
}

