% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reverse_ns.r
\name{reverse_ns}
\alias{reverse_ns}
\title{Reverse Nameserver}
\usage{
reverse_ns(domain, limit = NULL)
}
\arguments{
\item{domain}{domain to perform the  query on}

\item{limit}{Limits the size of the domain list than can appear in a
response.}
}
\value{
a \code{list} of result detais for the \code{domain}
}
\description{
The Reverse Name Server API provides a list of domain names that share the
same primary or secondary name server. You can request a name server
directly, or you can provide a domain name and the API will provide the
list of domain names pointed to the same name servers as those listed as
the primary and secondary name servers on the domain name you requested.
}
\examples{
reverse_ns("domaintools.net")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/reverse-name-server/}
}

