% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reverse_ip.R
\name{shared_ips}
\alias{shared_ips}
\title{Shared IPs}
\usage{
shared_ips(hostname, limit = NULL)
}
\arguments{
\item{hostname}{IP address to perform the reverse IP query}

\item{limit}{Limits the size of the domain list than can appear in a
response. The limit is applied per-IP address, not for the entire
request.}
}
\value{
a \code{list} of result detais for the \code{hostname}
}
\description{
The Reverse IP API provides a list of domain names that share the same
Internet host (i.e. the same IP address).
}
\note{
In rare cases, you may request an IP for which no recent Whois
      record is available. If that occurs, the system will respond with
      an error.
}
\examples{
shared_ips("domaintools.com")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/reverse-ip/}
}

