% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whois.r
\name{whois}
\alias{whois}
\title{Whois Lookup}
\usage{
whois(domain)
}
\arguments{
\item{domain}{domain to perform the Whois query on}
}
\value{
a \code{list} of basic Whois result detais for the \code{domain}
}
\description{
The Whois Lookup API provides the ownership record for a domain name or IP
address with basic registration details. The API is optimized to respond
quickly and is designed to handle a high volume of parallel requests. This
is the ideal product to use if you have a busy web site or a long list of
domains that you need to process.
}
\details{
The response includes the most recent Whois record we have for the domain
name or IP address you provided in the URL. This ensures a consistently fast
response without a dependency on external servers and usually returns a
record updated within the last few days. You can use the date property in
the whois response node to check the age of the record and confirm it was
updated recently enough for your needs.
}
\note{
In rare cases, you may request a domain for which no recent Whois
      record is available. If that occurs, the system will respond with
      an error.
}
\examples{
whois("domaintools.com")
whois("dailychanges.com")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/whois-lookup/}
}

