% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whois_history.r
\name{whois_history}
\alias{whois_history}
\title{Whois History}
\usage{
whois_history(domain)
}
\arguments{
\item{domain}{domain to perform the Whois query on}
}
\value{
a \code{list} of Whois history result detais for the \code{domain}
}
\description{
The Whois History API provides a list of historic Whois records for a
domain name. The format for each record is similar to that returned by
the Recent Whois API although some fields may be missing if the data was
not available when the record was stored in our system.
}
\note{
In rare cases, you may request a domain for which no recent Whois
      record is available. If that occurs, the system will respond with
      an error.
}
\examples{
whois_history("domaintools.com")
whois_history("dailychanges.com")
}
\references{
\url{http://www.domaintools.com/resources/api-documentation/whois-history/}
}

